requirements.txt
livekit-agents[google]~=1.5
python-dotenv~=1.0
playwright~=1.52
requests~=2.32
PyMuPDF~=1.26
pytest~=8.3
